How to use @State property wrapper in SwiftUI | Bootcamp #19
In this video we will learn how to properly use the @State property wrapper in SwiftUI. By adding @State to a variable, we can let the View know to watch the value of that variable so that if it changes, the View knows to update accordingly. State variables are used all the time in SwiftUI to create dynamic views that can change and animate while on the screen!