How to Extract Subviews in SwiftUI | Bootcamp #21
In this video we will learn how to extract subviews from our SwiftUI code. This is a very common practice that allows us to create reusable subviews that can have custom initializers. If you ever have an item in a view that will be repeated, but each time it requires different data (such as different text or colors), then you can make that item it's own subview.