How to use if-else and conditional statements in SwiftUI | Bootcamp #23
In this video we will learn how to use conditional logic in our SwiftUI applications.
We will first learn how to create "if - else" statements and then learn more advanced operations, such as "if-not", "or" & "and" statements.
This logic will help us create complex and dynamic views when we start building iOS applications!