How to use Ternary Operators in SwiftUI | Bootcamp #24
In this video we learn how to use ternary operators in our SwiftUI applications.
They are basically "if-else" statements that are condensed into one line of code.
Ternary operators are extremely powerful in all iOS applications and we can leverage them in SwiftUI to easily create complex views and animations.