How to use Transition in SwiftUI | Bootcamp #27
Transitions can be used to animation an object on-to or off-of the screen. Since the transition only occurs when the object is being added or removed from the view hierarchy, it can be a little tricky to implement correctly. In this video we will learn how to implement transitions and then explore several of the most popular animations for them.