How to use NavigationView and NavigationLink in SwiftUI | Bootcamp #30
The NavigationView is one of the most common containers in iOS Development. By adding our View inside a NavigationView, we get access to a lot of handy features, such as the Navigation Title, the Navigation Bar (with Navigation Bar Buttons and Toolbar Items) and the Navigation Link. In this video we will learn how to implement a NavigationView() and how to segue using the NavigationLink().