22. How to use ContextMenu in SwiftUI-BCMP-#34
We will learn how to use and implement a .contextMenu into our SwiftUI application. The Context Menu is a relatively new features in iOS that presents a menu of buttons to our user on top of an object. While the ContextMenu is not as popular as some of the other components we've learned so far (ie. the Alert and ActionSheet which we covered in the previous two videos), it is becoming increasingly more popular.