Code Update : Update to the property wrapper StateObject
StateObject and ObservedObject both are the same to update the view if any changes in the viewModel class.
However if the View changes that ViewModel class is also reloaded. So to avoid this the StateObject property wrapper is assigned where the ViewObject is initialised for the first time.
Whenever we are passing the viewModel to subsequent screen then we use the ObservedObject property wrapper.